refactor: extract GamesList component to new file
