+++ src/views/components/GamePageList/GamePageList.tsx
@@ -1,12 +1,11 @@
 import { IGamePageList } from './IGamePageList.ts'
 import style from './GamePageList.module.scss'
-import { CountFilter, GameCard, PageSelection } from '@views/components'
-import React, { ComponentProps, useState } from 'react'
+import { CountFilter, GamesList, PageSelection } from '@views/components'
+import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import {
 	GamePageListViewModel
 } from '@src/view-models/GamePageListViewModel.ts'
-import { Game } from '@data/types'
 
 
 export { GamePageList }
@@ -45,20 +44,3 @@ function GamePageList({ games }: IGamePageList) {
 	)
 }
 
-interface IGamesList extends ComponentProps<'ul'> {
-	currentGames: Game[]
-	onHandleClick(e: React.MouseEvent): void
-}
-
-function GamesList({ currentGames, onHandleClick }: IGamesList) {
-	return (
-		<ul className={style.GameList} onClick={onHandleClick}>
-			{currentGames.map(game => (
-				<li key={`game-${game.id}`}>
-					<GameCard game={game}/>
-				</li>
-			))
-			}
-		</ul>
-	)
-}
