+++ src/views/components/index.ts
@@ -21,3 +21,4 @@ export {
 	PageSelection
 } from '@views/components/PageSelection/PageSelection.tsx'
 export { CountFilter } from '@views/components/CountFilter/CountFilter.tsx'
+export { GamesList } from '@views/components/GamesList/GamesList.tsx'
